refactor: 패캐지 분리
