+++ app/src/main/java/nextstep/jwp/infrastructure/processor/RequestProcessor.java
@@ -1,6 +1,6 @@
 package nextstep.jwp.infrastructure.processor;
 
-import nextstep.jwp.model.CustomHttpRequest;
+import nextstep.jwp.model.web.request.CustomHttpRequest;
 
 import java.io.OutputStream;
