+++ app/src/main/java/nextstep/jwp/infrastructure/processor/ResourceRequestProcessor.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.infrastructure.processor;
 
-import nextstep.jwp.model.ContentType;
-import nextstep.jwp.model.CustomHttpRequest;
-import nextstep.jwp.model.CustomHttpResponse;
+import nextstep.jwp.model.web.ContentType;
+import nextstep.jwp.model.web.request.CustomHttpRequest;
+import nextstep.jwp.model.web.response.CustomHttpResponse;
 
 import java.io.File;
 import java.io.IOException;
