+++ app/src/main/java/nextstep/jwp/model/web/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.model;
+package nextstep.jwp.model.web;
 
 public enum ContentType {
     HTML("text/html", ".html"),
